- updated tests

+++ src/test/java/com/bankofcli/service/TransactionServiceIT.java
@@ -36,13 +36,13 @@ public class TransactionServiceIT {
     @AfterEach
     public void teardown(){
         DatabaseManager db = new DatabaseManager();
-        String deleteAccQry = "DELETE FROM accounts";
+        String deleteAccQry = "DELETE FROM accounts WHERE account_id <= 0";
         String deleteTransQry = "DELETE FROM transactions";
         try(Connection con = db.open();
             PreparedStatement psAcc = con.prepareStatement(deleteAccQry);
             PreparedStatement psTrans = con.prepareStatement(deleteTransQry)) {
-            psAcc.executeUpdate();
             psTrans.executeUpdate();
+            psAcc.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
